REMOVED update location form below the map
- will implement a separate form to allow modifications to the whole
response
- moved the html header to its own file so that it can be reused.
Currently it is being used by map and response. Edit Response will
also be using the same header
- changed the table resource to not hold custom field labels - deleted
those fields. On the other hand I added a field to store the map
name - This name will be used for the implementation of viewing all
the maps on one interface.