Fixed forms clearing when an error occurs during submission
Currently, images can be replaced by uploading a new image, but cannot
be deleted. Also since the file name for the image is the same, I had
to append a random number to the end for it to update on the interface.
(Note: the random number is not saved in the database)