REFACTORED image file name
I have appended a timestamp to the end of the filename. This is to
allow for the future implementation of multiple submissions per
user for a map

If an image has been uploaded before and the user is updating the
image, the filename stays the same. This is so that the "old image"
will be replaced by the new image and we would not need to deal with
removing "old images".
images that were uploaded before (due to same name)