FIX php8 upgrade issues
We needed to upgrade from php5 to php8 as 8 is the only version still in
support.

php7 fixes:
- Magic quotes are deprecated and removed from the language.
- gd's with-jpeg-dir config is no longer valid.
- UploadHandler needed an explicit cast from string to int to handle
  filesizes correctly.

Noticed that the function used to calculate filesize by byte is
duplicated in a lot of places when add the int cast fix to it. If we're
going to keep using this long term, we should probably dedupe it.

php8 fixes:
- Accessing undeclared class properties
- Undefined array keys

edit_response.php didn't have <body> tags for some reason. When adding
the body tag, I noticed that if you launch into responsemap in a new
tab, the elements go right up to the edge of the browser window, so I
added a padding to all the body tags to be a bit more user friendly.