FIX not working in canvas
The user ID sent by Canvas must have gotten longer or something, cause
it became too long to be stored in the database. I've increased the
length of all VARCHAR columns to 255 in order to fix this.

Debugging was hampered by the lack of errors, especially from mysqli, so
I've added a switch to turn on error display when the env var DEBUG is
set.

lti.php usedummydata() seems to be a typo.