FIX word cloud not working
Parsing of query params was using the php built-in split(), which was
removed some time ago. Looking at the parse_parameters() function, I
think it's functionally identical to the php built-in parse_str(), so
I've just replaced the entire function with that in OAuth.php.

The params to enable word cloud were not being set to $_SESSION, added
them in lti.php

The usecolor param should be read from $_SESSION instead of $_POST. This
might be a difference between how edX and Canvas passes these custom
params. The easiest way to pass this for Canvas is via GET queries added
on to the LTI launch, not via POST.

The word cloud was showing words with random <> attached for some
reason. Turns out there was a remove common words processing which
removed the p in <p> tags. I've fixed this by stripping all tags before
sending the word list to removeCommonWords().

Note, there's a caching issue with usecolor, you might need to clear
cache so the updated javascript for usecolor gets picked up.